Add simple/inefficient colliders to enemies

+++ include/headers/perlin_noise.h
@@ -3,18 +3,18 @@
 #include <cmath>
 #include <random>
 
-using vector2 = std::pair<double, double>;
+using vector2 = std::pair<float, float>;
 class PerlinNoise
 {
 public:
     PerlinNoise();
     PerlinNoise(unsigned _seed); 
     ~PerlinNoise();
-    double GetPerlinNoise2D(double x, double y); // With fractal brownian motion
+    float GetPerlinNoise2D(float x, float y); // With fractal brownian motion
 private:
-    double DotProduct(const vector2 &V1, const vector2 &V2);
+    float DotProduct(const vector2 &V1, const vector2 &V2);
     vector2 RandomGradient(int ix, int iy);
-    double Interpolate(double t, double a1, double a2);
-    double GetRawPerlinNoise2D(double x, double y);
+    float Interpolate(float t, float a1, float a2);
+    float GetRawPerlinNoise2D(float x, float y);
     unsigned seed;
 };
