+++ src/perlin_noise.cpp
@@ -11,7 +11,7 @@ PerlinNoise::PerlinNoise(unsigned _seed)
 }
 PerlinNoise::~PerlinNoise() {}
 
-double PerlinNoise::DotProduct(const vector2 &V1, const vector2 &V2)
+float PerlinNoise::DotProduct(const vector2 &V1, const vector2 &V2)
 {
     return V1.first * V2.first + V1.second * V2.second;
 }
@@ -33,7 +33,7 @@ vector2 PerlinNoise::RandomGradient(int ix, int iy) {
 }
 
 
-double PerlinNoise::Interpolate(double w, double a0, double a1)
+float PerlinNoise::Interpolate(float w, float a0, float a1)
 {
     // return a0 + (a1 - a0) * w;
 
@@ -44,50 +44,50 @@ double PerlinNoise::Interpolate(double w, double a0, double a1)
     return (a1 - a0) * ((w * (w * 6.0 - 15.0) + 10.0) * w * w * w) + a0;
 }
 
-double PerlinNoise::GetRawPerlinNoise2D(double x, double y)
+float PerlinNoise::GetRawPerlinNoise2D(float x, float y)
 {
     int X = ((int) floor(x));
     int Y = ((int) floor(y));
 
-    double xf = x - (double) X;
-    double yf = y - (double) Y;
+    float xf = x - (float) X;
+    float yf = y - (float) Y;
 
     vector2 topRight = std::make_pair(xf - 1.0, yf - 1.0);
     vector2 topLeft = std::make_pair(xf, yf - 1.0);
     vector2 bottomRight = std::make_pair(xf - 1.0, yf);
     vector2 bottomLeft = std::make_pair(xf, yf);
 
-    double dotTopRight = DotProduct(topRight, RandomGradient(X + 1, Y + 1));
-    double dotTopLeft = DotProduct(topLeft, RandomGradient(X, Y + 1));
-    double dotBottomRight = DotProduct(bottomRight, RandomGradient(X, Y + 1));
-    double dotBottomLeft = DotProduct(bottomLeft, RandomGradient(X, Y));
+    float dotTopRight = DotProduct(topRight, RandomGradient(X + 1, Y + 1));
+    float dotTopLeft = DotProduct(topLeft, RandomGradient(X, Y + 1));
+    float dotBottomRight = DotProduct(bottomRight, RandomGradient(X, Y + 1));
+    float dotBottomLeft = DotProduct(bottomLeft, RandomGradient(X, Y));
 
-    double u = xf;
-    double v = yf;
+    float u = xf;
+    float v = yf;
 
     return Interpolate(u, 
                         Interpolate(v, dotBottomLeft, dotTopLeft), 
                         Interpolate(v, dotBottomRight, dotTopRight));
     // return value approximately in [-1.0, 1.0]
-    // due to double imprecision
+    // due to float imprecision
 }
 
-double PerlinNoise::GetPerlinNoise2D(double x, double y)
+float PerlinNoise::GetPerlinNoise2D(float x, float y)
 {
-    double ret = 0;
-    double magnitude = 2;
-    double frequency = 0.08;
+    float ret = 0;
+    float magnitude = 2;
+    float frequency = 0.08;
     // With fractal brownian motion - only 2 here - increase for terrain complexity
     for (int o = 0; o < 2; o++)
     {
-        double tmp = magnitude * GetRawPerlinNoise2D(x * frequency, y * frequency);
+        float tmp = magnitude * GetRawPerlinNoise2D(x * frequency, y * frequency);
         ret += tmp;
         magnitude /= 2;
         frequency += frequency;
     }
 
     // scale from [-1, 1] to [0, 1]
-    // Just approximately in that range (due to double imprecision)
+    // Just approximately in that range (due to float imprecision)
     ret += 1.0;
     ret /= 2;
